requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
urllib3==2.0.5
fake-useragent==1.2.1
selenium==4.12.0
PyYAML==6.0.1
Flask==2.3.3
flask-restx==1.1.0
python-dotenv==1.0.0
selenium-stealth==1.0.6